backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.3
psycopg2-binary==2.9.10
dj-database-url==2.2.0
